Cygwin fixes by Ben van Klinken
For fun i tried compiling strigi with cygwin. I got the core things
working (deepfind, etc), here's a small patch of the things i had to
change to
make it work (against the current svn).
Mostly compile stuff (assumptions are made that if the compiler is not
msvc then it is mingw). Also there are some ambiguous casts for
addValue. And
a small typo in the strigiconfig.h.cmake file (which wasn't actually
necessary, but is probably incorrect).


svn path=/trunk/kdesupport/strigi/; revision=704666

+++ src/streamanalyzer/endanalyzers/mpegendanalyzer.cpp
@@ -85,8 +85,8 @@ char MpegEndAnalyzer::analyze(AnalysisResult& idx, InputStream* in) {
     std::map<std::string, const Strigi::RegisteredField*>tempfields
         = factory->fields;
     idx.addValue(tempfields["frame rate"], this->frame_rate);
-    idx.addValue(tempfields["dimensions.y"], this->vertical_size);
-    idx.addValue(tempfields["dimensions.x"], this->horizontal_size);
+    idx.addValue(tempfields["dimensions.y"], (int32_t)this->vertical_size);
+    idx.addValue(tempfields["dimensions.x"], (int32_t)this->horizontal_size);
     
     if (this->mpeg_version == 1) {
         idx.addValue(tempfields["video codec"], "MPEG-1");
